lily-model: Added IId<K>. Implements ITextParametric on entity types.

+++ lily/layer-2/lily-model/src/test/java/user/GroupPlayer.java
@@ -24,7 +24,7 @@ public class GroupPlayer
         Group group = new Group();
         group.setCodeName("Test" + random.nextInt());
         group.setLabel("Test1");
-        int id = groupMapper.insert(group);
+        int id = (int) groupMapper.insert(group);
 
         group.setId(id);
         group.setLabel("Changed");
